[fix] changed clangd compile flags for intellisense and add it to makefile
