+++ .clang-format
@@ -1,1 +1,2 @@
 BasedOnStyle: Google
+NamespaceIndentation: All
