+++ Makefile
@@ -1,5 +1,5 @@
 raycer:
-	g++ src/ray_main.cc -o raycer -O3 -fopenmp -march=native -mavx
+	g++ src/ray_main.cc -o raycer -Iinclude -O3 -fopenmp -march=native -mavx 
 
 clean:
 	rm raycer*
