+++ include/ray_def.hh
@@ -7,11 +7,8 @@
 #include <time.h>
 #include <xmmintrin.h>
 
-
 namespace raycer {
-
-typedef float
-    fnum;  // change the definition of num to gain higher precision, but
-           // this is not necessary to use double since float is enough
-
+  typedef float
+      fnum;  // change the definition of num to gain higher precision, but
+             // this is not necessary to use double since float is enough
 }  // namespace raycer
