--- src/ray_util.h
@@ -1,9 +1,0 @@
-#pragma once
-
-// include all the header of /util so that your compiler could search for the
-// header file
-
-#include "../util/ray_def.h"
-#include "../util/ray_image.h"
-#include "../util/ray_math.h"
-#include "../util/ray_misc.h"
