+++ src/raycer.hh
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "ray_def.hh"
+#include "ray_image.hh"
+#include "ray_math.hh"
+#include "ray_misc.hh"
